# docs/swagger.yaml
basePath: /api/v1
definitions:
  attributevolunteer.CreateRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    type: object
  attributevolunteer.UpdateRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      profile_photo:
        type: string
    required:
    - email
    - name
    - password
    type: object
  categoryactivity.CreateRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    type: object
  categoryactivity.UpdateRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
  description: API documentation
  title: Social BE API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /master/attribute-volunteers:
    get:
      description: Get paginated list of attribute volunteer
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all attribute volunteer
      tags:
      - attribute-volunteer
    post:
      consumes:
      - application/json
      description: Create new attribute volunteer
      parameters:
      - description: Create attribute volunteer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/attributevolunteer.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create attribute volunteer
      tags:
      - attribute-volunteer
  /master/attribute-volunteers/{id}:
    delete:
      description: Soft delete attribute volunteer
      parameters:
      - description: Attribute Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete attribute volunteer
      tags:
      - attribute-volunteer
    get:
      description: Get detail attribute volunteer
      parameters:
      - description: Attribute Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get attribute volunteer by ID
      tags:
      - attribute-volunteer
    put:
      consumes:
      - application/json
      description: Update existing attribute volunteer
      parameters:
      - description: Attribute Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update attribute volunteer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/attributevolunteer.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update attribute volunteer
      tags:
      - attribute-volunteer
  /master/category-activities:
    get:
      description: Get paginated list of category activity
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all category activity
      tags:
      - category-activity
    post:
      consumes:
      - application/json
      description: Create new category activity
      parameters:
      - description: Create category activity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/categoryactivity.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create category activity
      tags:
      - category-activity
  /master/category-activities/{id}:
    delete:
      description: Soft delete category activity
      parameters:
      - description: Category Activity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete category activity
      tags:
      - category-activity
    get:
      description: Get detail category activity
      parameters:
      - description: Category Activity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get category activity by ID
      tags:
      - category-activity
    put:
      consumes:
      - application/json
      description: Update existing category activity
      parameters:
      - description: Category Activity ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update category activity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/categoryactivity.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update category activity
      tags:
      - category-activity
...
  /register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"